% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_dl_models.R
\name{download_dl_models}
\alias{download_dl_models}
\title{Download deep learning models for Greenland halibut age estimation}
\usage{
download_dl_models(
  data_storage_path = "~/AgeEstimatoR large files",
  download_path = "https://owncloud.imr.no/index.php/s/9xDhXY5OZZzTI0P/download"
)
}
\arguments{
\item{data_storage_path}{Character defining the folder where the deep learning models should be downloaded.}

\item{download_path}{Character defining the URL or server folder where the deep learning models should be downloaded from. The default value indicates the currently recommended location.}
}
\value{
Returns a character file path to the folder containing the deep learning models. Can also be used in cases when the folder exists.
}
\description{
The function downloads the deep learning models required by the \code{\link{estimate_age}}.
}
\details{
The function downloads the pre-existing deep learning models required for Greenland halibut age estimation.
}
\author{
Mikko Vihtakari
}
