% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_age.R
\name{estimate_age}
\alias{estimate_age}
\title{Estimate age from standardized otolith images using deep learning}
\usage{
estimate_age(
  image_path,
  model_path = "~/AgeEstimatoR large files/dl_models",
  output_path = NULL,
  venv_path = "./python_virtualenv"
)
}
\arguments{
\item{image_path}{Character defining the file path to the standardized images from \code{\link{standardize_images}}.}

\item{model_path}{Character defining the file path to the deep learning models by Martinsen et al. (2022). If the specified folder does not exist, the \code{\link{download_dl_models}} function is used to download the models.}

\item{output_path}{Character defining the file path to the model output containing file name without extension. If \code{NULL}, the data are not saved to a file, but returned as a data frame instead.}

\item{venv_path}{Character defining path to the Python virtual environment containing the TensorFlow package. If the virtual environment does not exist, it will be created and required packages downloaded.}
}
\value{
Saves a jpg image in \code{output_path}. The file name is similar to the original images except the addition of \code{prefix} and \code{postfix}.
}
\description{
Estimates age from standardized otolith images (see \code{\link{standardize_images}}) using deep learning algorithms developed for Greenland halibut by Martinsen et al. (2022).
}
\author{
Iver Martinsen, Mikko Vihtakari, Alf Harbitz, Tine Nilsen (Institute of Marine Research, Norway)
}
