% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_python.R
\name{setup_python}
\alias{setup_python}
\title{Set up Python virtual environment and download required modules}
\usage{
setup_python(venv_path = "~/AgeEstimatoR large files/python_virtualenv")
}
\arguments{
\item{venv_path}{Character defining path to the Python virtual environment containing the TensorFlow package. If the virtual environment does not exist, it will be created and required packages downloaded.}
}
\description{
The function sets up Python and TensorFlow as required by the \code{\link{estimate_age}}.
}
\details{
The function locates the Python and TensorFlow installation in the python_virtualenv folder. If the folder does not exist, the function creates the folder with a Python virtual environment where these packages will be installed. The folder takes 1.43 Gb of file space.
}
\author{
Mikko Vihtakari
}
