% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardize_image.R
\name{standardize_image}
\alias{standardize_image}
\title{Standardize a single otolith image for age estimation by deep learning}
\usage{
standardize_image(
  imno,
  input_path,
  output_path,
  adj = 1,
  fillparam = 9,
  prefix = sprintf("\%d_", imno),
  postfix = ".jpg",
  ask = TRUE
)
}
\arguments{
\item{imno}{Integer giving the otolith number, i.e. the file serial number in \code{list.files(input_path)}.}

\item{input_path}{Character defining the file path to input images.}

\item{output_path}{Character defining the file path where the processed images should be saved.}

\item{adj}{Numeric defining the adjustment factor for intensity threshold (the \code{adjust} parameter in \code{\link[imager]{threshold}}).}

\item{fillparam}{Integer defining the morphological filter size used to fill holes within the otolith contour. Must be an odd number. See \code{\link[imager]{fill}}.}

\item{prefix}{Character or function defining the prefix for the image file. Modifications here might mess up the \code{\link{estimate_age}} function.}

\item{postfix}{Character or function defining the postfix for the image file. Must contain file extension, which defines the file format.}

\item{ask}{Logical indicating whether the function should ask for new threshold (\code{adj}) and fill (\code{fillparam}) parameters when the user is not satisfied with the image quality.}
}
\value{
Saves a jpg image in \code{output_path}. The file name is similar to the original images except the addition of \code{prefix} and \code{postfix}.
}
\description{
Standardizes otolith images ready for age estimation using the \code{\link{estimate_age}} function. Creates a 256 X 256 X 3 right otolith image with manual framing.
}
\author{
Tine Nilsen, Mikko Vihtakari, Kristin Windsland, Alf Harbitz (Institute of Marine Research, Norway)
}
